requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.8.0
argcomplete==3.6.0
async-timeout==4.0.3
attrs==25.1.0
beautifulsoup4==4.13.3
boto3==1.37.9
botocore==1.37.9
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cohere==5.14.0
colorama==0.4.6
Crawl4AI==0.5.0.post4
cryptography==44.0.2
cssselect==1.2.0
Deprecated==1.2.18
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
dotenv==0.9.9
email_validator==2.2.0
eval_type_backport==0.2.2
exceptiongroup==1.2.2
fake-http-header==0.3.5
fake-useragent==2.0.3
fastapi==0.115.11
fastapi-cli==0.0.7
fastavro==1.10.0
faust-cchardet==2.1.19
filelock==3.17.0
frozenlist==1.5.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.38.0
gotrue==2.11.4
greenlet==3.1.1
griffe==1.6.0
groq==0.18.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.2
humanize==4.12.1
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.5.2
Jinja2==3.1.6
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.3.20
langchain-core==0.3.43
langchain-openai==0.3.8
langchain-text-splitters==0.3.6
langsmith==0.3.13
litellm==1.63.3
logfire-api==3.7.1
lxml==5.3.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mistralai==1.5.1
multidict==6.1.0
mypy-extensions==1.0.0
nltk==3.9.1
numpy==2.0.2
openai==1.65.4
opentelemetry-api==1.30.0
orjson==3.10.15
packaging==24.2
pillow==10.4.0
playwright==1.50.0
postgrest==0.19.3
prompt_toolkit==3.0.50
propcache==0.3.0
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic-ai==0.0.36
pydantic-ai-slim==0.0.36
pydantic-graph==0.0.36
pydantic_core==2.27.2
pyee==12.1.1
Pygments==2.19.1
pyOpenSSL==25.0.0
pyperclip==1.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-git==2018.2.1
python-multipart==0.0.20
PyYAML==6.0.2
rank-bm25==0.2.2
realtime==2.4.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rich-toolkit==0.13.2
rpds-py==0.23.1
rsa==4.9
s3transfer==0.11.4
Send2Trash==1.8.3
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.6
SQLAlchemy==2.0.38
starlette==0.46.0
storage3==0.11.3
StrEnum==0.4.15
supabase==2.13.0
supafunc==0.9.3
tenacity==9.0.0
tf-playwright-stealth==1.1.2
tiktoken==0.9.0
tokenizers==0.21.0
tqdm==4.67.1
typer==0.15.2
types-requests==2.31.0.6
types-urllib3==1.26.25.14
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.12.2
urllib3==1.26.20
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
wcwidth==0.2.13
websockets==14.2
wrapt==1.17.2
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
